Move LoginForm components to forms directory
Relocated LoginForm.tsx and LoginForm.module.scss to a new 'forms' subdirectory for better organization. Updated all relevant import paths in App.tsx, RegisterForm.module.scss, and LoginForm.tsx to reflect the new structure.
