+++ printlib-frontend/src/components/RegisterForm.module.scss
@@ -1,5 +1,5 @@
 @import '../styles/variables';
-@import './LoginForm.module.scss';
+@import './forms/LoginForm.module.scss';
 
 .buttonRow {
     display: flex;
