+++ printlib-frontend/src/components/forms/LoginForm.module.scss
@@ -1,4 +1,4 @@
-@import '../styles/variables';
+@import '../../styles/variables';
 
 .loginForm {
   background-color: var(--color-background);
